completed advanced calculator and passed all tests

+++ src/problems/advancedcalculator/Multiplication.java
@@ -2,8 +2,14 @@ package problems.advancedcalculator;
 
 public class Multiplication {
 
-    static String performMultiplication(int[] numbers) {
-        int multiplication = numbers[0] * numbers[1];
+//    static String performMultiplication(int[] numbers) {
+//        int multiplication = numbers[0] * numbers[1];
+//        System.out.println(" Multiplication " + multiplication);
+//        return String.valueOf(multiplication);
+//    }
+
+    static String performMultiplication(double[] numbers) {
+        double multiplication = numbers[0] * numbers[1];
         System.out.println(" Multiplication " + multiplication);
         return String.valueOf(multiplication);
     }
