+++ src/problems/advancedcalculator/Subtraction.java
@@ -1,8 +1,14 @@
 package problems.advancedcalculator;
 
 public class Subtraction {
-    static String performSubtraction(int[] numbers) {
-        int subtraction = numbers[0] - numbers[1];
+//    static String performSubtraction(int[] numbers) {
+//        int subtraction = numbers[0] - numbers[1];
+//        System.out.println("Subraction is " + subtraction);
+//        return String.valueOf(subtraction);
+//    }
+
+    static String performSubtraction(double[] numbers) {
+        double subtraction = numbers[0] - numbers[1];
         System.out.println("Subraction is " + subtraction);
         return String.valueOf(subtraction);
     }
